fix update to 25.10.10 (#9917)
